requirements.txt
alembic==1.12.1
bcrypt==4.2.1
certifi==2024.12.14
cffi==2.0.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==46.0.3
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.2.5
Flask-Bcrypt==1.0.1
Flask-Dance==7.1.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==3.1.1
idna==3.10
importlib-metadata==6.7.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.4
Mako==1.2.4
MarkupSafe==2.1.5
oauthlib==3.2.2
Pillow==9.5.0
pycparser==2.23
PyMySQL==1.1.1
python-dotenv==1.2.1
requests==2.31.0
requests-oauthlib==2.0.0
SQLAlchemy==2.0.36
typing_extensions==4.7.1
urllib3==2.0.7
URLObject==2.4.3
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.15.0
